+++ RSS-CSS-Selectors/src/index.js
@@ -1,5 +1,3 @@
 import "./index.scss";
 import "./index.html";
 import "./main.ts";
-import "./img/github.png";
-import "./img/rsschool.svg";
