+++ RSS-CSS-Selectors/src/modules/selectors.ts
@@ -1,0 +1,81 @@
+export function getMenuItems() {
+    return document.querySelectorAll('.menu-right h1');
+  }
+  
+  export function getTable() {
+    return document.querySelector('.table') as HTMLTableElement;
+  }
+  
+  export function getTableWrapper() {
+    return document.querySelector('.table-wrapper') as HTMLTableElement;
+  }
+  
+  export function getWindowModal() {
+    return document.querySelector('.window') as HTMLElement;
+  }
+  
+  export function getInputWindow() {
+    return document.querySelector('.input-window__markup') as HTMLElement;
+  }
+  
+  export function getInput() {
+    return document.querySelector('.input-window__strobe') as HTMLInputElement;
+  }
+  
+  export function getEditor() {
+    return document.querySelector('.editor') as HTMLInputElement;
+  }
+  
+  export function getButton() {
+    return document.querySelector('.button-help') as HTMLButtonElement;
+  }
+  
+  export function getTableStrobe() {
+    return document.querySelectorAll('.table') as NodeListOf<HTMLElement>;
+  }
+  
+  export function getTableWindow() {
+    return document.querySelector('.window') as HTMLElement;
+  }
+  
+  export function getTargetElementOne() {
+    return document.getElementById('one');
+  }
+  
+  export function getTargetElementTwo() {
+    return document.getElementById('two');
+  }
+  
+  export function getTargetElementThree() {
+    return document.getElementById('three');
+  }
+  
+  export function getTargetElementFour() {
+    return document.getElementById('four');
+  }
+  
+  export function getTargetElementFive() {
+    return document.getElementById('five');
+  }
+  
+  export function getTargetElementSix() {
+    return document.getElementById('six');
+  }
+  
+  export function getTargetElementSeven() {
+    return document.getElementById('seven');
+  }
+  
+  export function getTargetElementEight() {
+    return document.getElementById('eight');
+  }
+  
+  export function getTargetElementNine() {
+    return document.getElementById('nine');
+  }
+  
+  export function getTargetElementTen() {
+    return document.getElementById('ten');
+  }
+  
+
